Reverting back the Scope class interface
