+++ jackson-jq/src/main/java/net/thisptr/jackson/jq/BuiltinFunctionLoader.java
@@ -17,7 +17,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.MappingIterator;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-
 import net.thisptr.jackson.jq.internal.IsolatedScopeQuery;
 import net.thisptr.jackson.jq.internal.JsonQueryFunction;
 import net.thisptr.jackson.jq.internal.javacc.ExpressionParser;
@@ -77,8 +76,8 @@ public class BuiltinFunctionLoader {
 	 */
 	public Map<String, Function> listFunctions(final ClassLoader classLoader, final Version version, final Scope closureScope) {
 		final Map<String, Function> functions = new HashMap<>();
-		loadMacros(functions, classLoader, version, closureScope);
-		loadBuiltinFunctions(functions, version, classLoader);
+		functions.putAll(loadFunctionsFromJsonJq(classLoader, version, closureScope));
+		functions.putAll(loadFunctionsFromServiceLoader(classLoader, version));
 		return functions;
 	}
 
@@ -143,7 +142,8 @@ public class BuiltinFunctionLoader {
 		return annotation.value();
 	}
 
-	public void loadBuiltinFunctions(final Map<String, Function> functions, final Version version, final ClassLoader classLoader) {
+	public Map<String, Function> loadFunctionsFromServiceLoader(final ClassLoader classLoader, final Version version) {
+		final Map<String, Function> functions = new HashMap<>();
 		for (final Function fn : ServiceLoader.load(Function.class, classLoader)) {
 			String[] names = extractFunctionNamesFromAnnotationIfVersionMatch(fn, version);
 			if (names == null) { // i.e. if annotation is missing,
@@ -157,10 +157,12 @@ public class BuiltinFunctionLoader {
 			for (final String name : names)
 				functions.put(name, fn);
 		}
+		return functions;
 	}
 
-	public void loadMacros(final Map<String, Function> functions, final ClassLoader classLoader, final Version version, final Scope closureScope) {
+	public Map<String, Function> loadFunctionsFromJsonJq(final ClassLoader classLoader, final Version version, final Scope closureScope) {
 		try {
+			final Map<String, Function> functions = new HashMap<>();
 			final List<JqJson> configs = loadConfig(classLoader, CONFIG_PATH);
 			for (final JqJson jqJson : configs) {
 				for (final JqJson.JqFuncDef def : jqJson.functions) {
@@ -169,6 +171,7 @@ public class BuiltinFunctionLoader {
 					functions.put(def.name + "/" + def.args.size(), new JsonQueryFunction(def.name, def.args, new IsolatedScopeQuery(ExpressionParser.compile(def.body, version)), closureScope));
 				}
 			}
+			return functions;
 		} catch (final IOException e) {
 			throw new RuntimeException("Failed to load macros", e);
 		}
