--- jackson-jq/src/main/java/net/thisptr/jackson/jq/ObjectMapperProvider.java
@@ -1,8 +1,0 @@
-package net.thisptr.jackson.jq;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-public interface ObjectMapperProvider {
-
-    ObjectMapper get();
-}
