+++ jackson-jq/src/main/java/net/thisptr/jackson/jq/Scope.java
@@ -1,43 +1,141 @@
 package net.thisptr.jackson.jq;
 
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.TreeMap;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+
+import net.thisptr.jackson.jq.internal.misc.JsonQueryJacksonModule;
 import net.thisptr.jackson.jq.path.Path;
 
-public interface Scope {
+public class Scope {
+    private static final ObjectMapper DEFAULT_MAPPER = new ObjectMapper()
+            .registerModule(JsonQueryJacksonModule.getInstance());
 
-    static Scope newEmptyScope() {
-        return new DefaultScope(null);
+    @JsonProperty("functions")
+    private Map<String, String> debugFunctions() {
+        final Map<String, String> result = new TreeMap<>();
+        for (final Entry<String, Function> f : functions.entrySet())
+            result.put(f.getKey(), f.getValue().toString());
+        return result;
     }
 
-    static Scope newChildScope(final Scope scope) {
-        return new DefaultScope(scope);
-    }
+    @JsonProperty("parent")
+    private Scope parentScope;
 
-    void addFunction(final String name, final int n, final Function q);
+    @JsonIgnore
+    private Map<String, Function> functions;
 
-    void addFunction(final String name, final Function q);
-
-    Function getFunction(final String name, final int nargs);
-
-    Function getFunctionRecursive(final String name);
-
-    void setValue(final String name, final JsonNode value);
-
-    void setValueWithPath(final String name, final JsonNode value, final Path path);
-
-    ValueWithPath getValueWithPath(final String name);
-
-    JsonNode getValue(final String name);
-
-    ObjectMapper getObjectMapper();
-
-    ObjectMapperProvider getObjectMapperProvider();
-
-    interface ValueWithPath {
-
+    public interface ValueWithPath {
         JsonNode value();
 
         Path path();
+    }
+
+    private static class ValueWithPathImpl implements ValueWithPath {
+        @JsonProperty("value")
+        private final JsonNode value;
+
+        @JsonProperty("path")
+        private final Path path;
+
+        public ValueWithPathImpl(final JsonNode value, final Path path) {
+            this.value = value;
+            this.path = path;
+        }
+
+        @Override
+        public JsonNode value() {
+            return value;
+        }
+
+        @Override
+        public Path path() {
+            return path;
+        }
+    }
+
+    @JsonProperty("variables")
+    private Map<String, ValueWithPath> values;
+
+    @JsonIgnore
+    private ObjectMapper mapper = DEFAULT_MAPPER;
+
+    private Scope(final Scope parentScope) {
+        this.parentScope = parentScope;
+    }
+
+    public static Scope newEmptyScope() {
+        return new Scope(null);
+    }
+
+    public static Scope newChildScope(final Scope scope) {
+        return new Scope(scope);
+    }
+
+    public void addFunction(final String name, final int n, final Function q) {
+        addFunction(name + "/" + n, q);
+    }
+
+    public void addFunction(final String name, final Function q) {
+        if (functions == null)
+            functions = new HashMap<>();
+        functions.put(name, q);
+    }
+
+    public Function getFunction(final String name, final int nargs) {
+        final Function f = getFunctionRecursive(name + "/" + nargs);
+        if (f != null)
+            return f;
+        return getFunctionRecursive(name);
+    }
+
+    private Function getFunctionRecursive(final String name) {
+        if (functions != null) {
+            final Function q = functions.get(name);
+            if (q != null)
+                return q;
+        }
+        if (parentScope == null)
+            return null;
+        return parentScope.getFunctionRecursive(name);
+    }
+
+    public void setValue(final String name, final JsonNode value) {
+        setValueWithPath(name, value, null);
+    }
+
+    public void setValueWithPath(final String name, final JsonNode value, final Path path) {
+        if (values == null)
+            values = new HashMap<>();
+        values.put(name, new ValueWithPathImpl(value, path));
+    }
+
+    public ValueWithPath getValueWithPath(final String name) {
+        if (values != null) {
+            final ValueWithPath value = values.get(name);
+            if (value != null)
+                return value;
+        }
+        if (parentScope == null)
+            return null;
+        return parentScope.getValueWithPath(name);
+    }
+
+    public JsonNode getValue(final String name) {
+        final ValueWithPath value = getValueWithPath(name);
+        if (value == null)
+            return null;
+        return value.value();
+    }
+
+    @JsonIgnore
+    public ObjectMapper getObjectMapper() {
+        return mapper;
     }
 }
